Initial commit
Going to use github to cloud save this this project. It's already about where it needs to be for the scope of the assignment. I have ideas of what my bachelor thesis will be about and this project may be useful at that time. Other than that, it's to fill my git with something as it looks rather empty atm.