QUICK FIX
Relocated logic and added a Thread.Sleep for each room built as the Random class couldn't manage the rate of room creation.
Th oved logic is focused on setting the tile type, which is now arranged and handled in the rommbuilder class.
Further testing is needed, however it now works in 90% of tests.

Next:
Instead of a single entrypoint for each room, add an outer bound surrounding the room.

Remove the Tile weights and configure the AStar pathing to exclude any diagonal moves.

Implement new weight system (Interpolation) after the previous changes above to create some  bends to the,now, otherwise straight pathing.

Future:
apply a height value ( or use weight value) and apply brighness or darkness based on value to create some depth.
(May have to write a simple shader for this)